[#18660] conda-package: Prevent searching for 'libdl.so.2' on macOS
Searching for 'libdl.so.2' and creating it's link was primarily intended
for Ubuntu but it doesn't make a problem for any GNU/Linux, as they all
contain that library.

It causes problems on macOS, however, because there is no such library,
so the 'find / ...' command is executed each time conda activates the
environment with the Renode package installed. Therefore, checking
whether the OS is 'Linux' was added as the first test.

The single '=' is intended. Not all shells support '==' in 'test'.

+++ tools/packaging/conda/activate.sh
@@ -2,6 +2,6 @@
 # to avoid renode crashing we add local link on first activation.
 # As we link to the link provided by libc package it should work after update
 
-if ( ! [ -e $CONDA_PREFIX/lib/libdl.so ] ) && [ `whereis libdl | grep 'libdl\.so' | wc -l` -eq 0 ] ; then
+if [ `uname` = 'Linux' ] && ( ! [ -e $CONDA_PREFIX/lib/libdl.so ] ) && [ `whereis libdl | grep 'libdl\.so' | wc -l` -eq 0 ] ; then
     find / -name 'libdl.so.2' -exec ln -s {} $CONDA_PREFIX/lib/libdl.so \;
 fi
